# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: ISCP API
  version: 0.1.0
  summary: API-first ISCP protocol v2 reference services
servers:
  - url: http://localhost:8080
    description: Relay Reference Service
  - url: http://localhost:8081
    description: Trust Root Reference Service
paths:
...
  /v2/relay/devices/register-with-ticket:
    post:
      summary: Register a device using a pairing ticket
      responses:
        "200": { description: Access and refresh credentials }
        "409": { description: Ticket already consumed }
  /v2/relay/devices/refresh-access:
    post:
      summary: Refresh relay access
      responses:
        "200": { description: New access and refresh credentials }
        "401": { description: Invalid refresh credential }
...